@@ -1,25 +1,29 @@
 {
-  "name": "zelle_app",
+  "name": "zelleapp",
   "version": "1.0.0",
-  "main": "node_modules/expo/AppEntry.js",
+  "main": "index.js",
   "scripts": {
-    "start": "expo start",
-    "android": "expo start --android",
-    "ios": "expo start --ios",
+    "start": "expo start --dev-client",
+    "android": "expo run:android",
+    "ios": "expo run:ios",
     "web": "expo start --web"
   },
   "dependencies": {
     "@react-native-firebase/app": "^16.4.6",
+    "@react-native-firebase/firestore": "^16.4.6",
+    "@react-native-firebase/functions": "^16.4.6",
     "@react-native-firebase/storage": "^16.4.6",
-    "expo": "~46.0.16",
-    "expo-status-bar": "~1.4.0",
-    "firebase": "^9.13.0",
-    "firebase-admin": "^11.2.0",
-    "react": "18.0.0",
-    "react-native": "0.69.6"
+    "expo": "^47.0.8",
+    "expo-splash-screen": "~0.17.5",
+    "expo-status-bar": "~1.4.2",
+    "firebase": "^9.14.0",
+    "firebase-admin": "^11.3.0",
+    "react": "18.1.0",
+    "react-native": "0.70.5"
   },
   "devDependencies": {
-    "@babel/core": "^7.12.9"
+    "@babel/core": "^7.12.9",
+    "react-native-clean-project": "^4.0.1"
   },
   "private": true
 }
